Cleanup v2

+++ Controllers/UserController.php
@@ -28,58 +28,39 @@ class UserController implements IController
     {
         $this->database = new Database();
     }
-
+    
     public function create(array $data): ?array
     {
-        try {
-            $this->database->create(self::$table, $data);
-            return $this->database->getByColumn(self::$table, 'email', $data['email']);
-        } catch (Error $error) {
-            throw new Error("Gebruiker kon niet aangemaakt worden!");
-        }
+        $this->database->create(self::$table, $data);
+        
+        return $this->database->getByColumn(self::$table, 'email', $data['email']);
     }
 
     public function get(int $id): ?array
     {
-        try {
-            return $this->database->get(self::$table, $id);
-        } catch (Error $error) {
-            throw new Error("Gebruiker, waarvan ID = $id, niet gevonden!");
-        }
+        return $this->database->get(self::$table, $id);
     }
 
     public function index(): ?array
     {
-        try {
-            return $this->database->index(self::$table);
-        } catch (Error $error) {
-            throw new Error("Geen gebruikers gevonden!");
-        }
+        return $this->database->index(self::$table);
     }
 
     public function update(int $id, array $data): ?array
     {
-        // First check if item exists.
-        $this->get($id);
-
-        try {
+        if ($this->get($id)) {
             $this->database->update(self::$table, $id, $data);
-            return $this->get($id);
-        } catch (Error $error) {
-            throw new Error("Gebruiker, waarvan ID = $id, niet geupdate!");
         }
+        
+        return $this->get($id) ?? null;
     }
 
     public function delete(int $id): ?array
     {
-        // First check if item exists.
-        $user = $this->get($id);
-
-        try {
+        if ($user = $this->get($id)) {
             $this->database->delete(self::$table, $id);
-            return $user;   // Return deleted user when success.
-        } catch (Error $error) {
-            throw new Error("Gebruiker, waarvan ID = $id, niet verwijderd!");
         }
+        
+        return $user ?? null;
     }
 }
